refactor: 공통 로직 TransactionCommonProcessor로 분리
