+++ src/main/java/org/c4marathon/assignment/application/TransactionProcessor.java
@@ -5,8 +5,6 @@ import org.c4marathon.assignment.domain.AccountRepository;
 import org.c4marathon.assignment.domain.User;
 import org.c4marathon.assignment.domain.UserRepository;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Isolation;
-import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 
@@ -16,15 +14,7 @@ public class TransactionProcessor {
 	private final AccountRepository accountRepository;
 	private final UserRepository userRepository;
 	private final WireTransferStrategyContext wireTransferStrategyContext;
-
-	@Transactional(isolation = Isolation.READ_COMMITTED)
-	public void updateBalance(String accountNumber, long money) {
-		if (money > 0) {
-			updateUser(accountNumber, money);
-		}
-
-		updateAccount(accountNumber, money);
-	}
+	private final TransactionCommonProcessor transactionCommonProcessor;
 
 	public void wireTransfer(String senderAccountNumber, String receiverAccountNumber, long money) {
 		Account senderAccount = accountRepository.findByAccountNumber(senderAccountNumber)
@@ -38,20 +28,7 @@ public class TransactionProcessor {
 		wireTransferStrategy.wireTransfer(senderAccountNumber, receiverAccountNumber, money);
 	}
 
-	private void updateUser(String accountNumber, long money) {
-		Account account = accountRepository.findByAccountNumber(accountNumber)
-			.orElseThrow(() -> new RuntimeException("Account not found"));
-
-		int updatedRow = userRepository.charge(account.getUserId(), money);
-
-		if (updatedRow == 0)
-			throw new RuntimeException("Account not charged");
-	}
-
-	private void updateAccount(String accountNumber, long money) {
-		int updatedRow = accountRepository.updateBalance(accountNumber, money);
-
-		if (updatedRow == 0)
-			throw new RuntimeException("Failed to update balance.");
+	public void updateBalance(String accountNumber, long money) {
+		transactionCommonProcessor.updateBalance(accountNumber, money);
 	}
 }
